ServiceDiscovery/requirements.txt
flask
flask_cors
flask_talisman
flask_restx
flask_caching
flask_limiter
requests_random_user_agent
bs4
pandas
numpy
lxml
PyGithub